+++ Source/Networking/MaskedDogData.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace PlayerDogModel_Plus.Source.Networking
+{
+    [Serializable]
+    internal class MaskedDogData
+    {
+        public ulong maskedEnemyNetworkId;
+        public ulong mimickingClientId;
+    }
+}