+++ Source/Util/DogModelConstraints.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine.Animations;
+
+namespace PlayerDogModel_Plus.Source.Util
+{
+    internal class DogModelConstraints
+    {
+        internal PositionConstraint torso;
+        internal RotationConstraint head;
+        internal RotationConstraint armL;
+        internal RotationConstraint armR;
+        internal RotationConstraint legL;
+        internal RotationConstraint legR;
+    }
+}